Add files via upload
GitHub’s preview for .ipynb files may not fully render the content. For the best viewing experience, download the HTML files and open them in a browser, or open the .dbc files directly in Databricks. Additionally, a PDF file with screenshots of the original Databricks notebooks is attached for reference.